@@ -9,7 +9,6 @@ import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class DisruptionServiceImpl implements DisruptionService {
@@ -22,8 +21,9 @@ public class DisruptionServiceImpl implements DisruptionService {
 
     @Override
     public Flux<Disruption> getDisruptions(List<DisturbanceType> types) {
-        return client.getDisturbances(types.stream().map(Mappings::map).collect(Collectors.toList()))
+        return client.getDisturbances(types.stream().map(Mappings::map).toList())
                 .filter(it -> it instanceof DisruptionWrapper)
-                .map(it -> Mappings.map((DisruptionWrapper) it));
+                .map(DisruptionWrapper.class::cast)
+                .map(Mappings::map);
     }
 }
