@@ -1,6 +1,5 @@
 package md.vnastasi.cloud;
 
-import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.context.ApplicationPidFileWriter;
@@ -11,7 +10,7 @@ import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
 public class EurekaServerApplication {
 
     public static void main(String[] args) {
-        SpringApplication application = new SpringApplicationBuilder(EurekaServerApplication.class).build();
+        var application = new SpringApplicationBuilder(EurekaServerApplication.class).build();
         application.addListeners(new ApplicationPidFileWriter());
         application.run(args);
     }
