@@ -6,23 +6,23 @@ import org.springframework.boot.context.properties.ConstructorBinding;
 @ConfigurationProperties(prefix = "application")
 @ConstructorBinding
 public record ApplicationProperties(
-    NsApiProperties nsApi
+        NsApiProperties nsApi
 ) {
 
-    public static record NsApiProperties(
-        String baseUrl,
-        PublicTravelInfoProperties publicTravelInfo
+    public record NsApiProperties(
+            String baseUrl,
+            PublicTravelInfoProperties publicTravelInfo
     ) {
 
-        public static record PublicTravelInfoProperties(
-            String basePath,
-            ApiKeyProperties apiKey,
-            String stationsPath
+        public record PublicTravelInfoProperties(
+                String basePath,
+                ApiKeyProperties apiKey,
+                String stationsPath
         ) {
 
-            public static record ApiKeyProperties(
-                String header,
-                String value
+            public record ApiKeyProperties(
+                    String header,
+                    String value
             ) {
             }
         }
