@@ -1,6 +1,5 @@
 package md.vnastasi.cloud;
 
-import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.context.ApplicationPidFileWriter;
@@ -13,7 +12,7 @@ import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 public class StationRegistryApplication {
 
     public static void main(String[] args) {
-        SpringApplication application = new SpringApplicationBuilder(StationRegistryApplication.class).build();
+        var application = new SpringApplicationBuilder(StationRegistryApplication.class).build();
         application.addListeners(new ApplicationPidFileWriter());
         application.run(args);
     }
