@@ -1,12 +1,11 @@
 package md.vnastasi.cloud.client.model;
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.List;
 
-@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
 public record StationsResponseWrapper(
-    @JsonProperty("payload") List<StationWrapper> payload
+        @JsonProperty("payload")
+        List<StationWrapper> payload
 ) {
 }
